Fix heading and message for sign message
