+++ src/components/SignMessageCompact.vue
@@ -5,15 +5,15 @@ import { useRoute } from 'vue-router'
 import { useAppStore } from '@/store/app'
 import { useParentConnectionStore } from '@/store/parentConnection'
 import { useRequestStore } from '@/store/request'
-import { useUserStore } from '@/store/user'
-import { methodAndAction } from '@/utils/method'
 
-defineProps({
-  request: {
-    type: Object,
-    required: true,
-  },
-})
+type SignMessageCompactProps = {
+  title: string
+  deprecated: boolean
+  permission: string
+  request: any
+}
+
+const props = defineProps<SignMessageCompactProps>()
 
 const emits = defineEmits(['reject', 'approve'])
 
@@ -21,81 +21,6 @@ const appStore = useAppStore()
 const route = useRoute()
 const requestStore = useRequestStore()
 const parentConnectionStore = useParentConnectionStore()
-const userStore = useUserStore()
-
-const stateChangeRequests = [
-  methodAndAction.wallet_addEthereumChain,
-  methodAndAction.wallet_switchEthereumChain,
-  methodAndAction.wallet_watchAsset,
-]
-
-function getTitle(requestMethod: string) {
-  if (stateChangeRequests.includes(requestMethod)) {
-    return requestMethod
-  }
-  return 'Sign Message'
-}
-
-function isSiweMessage(message: string) {
-  return (
-    message.includes('wants you to sign in with your Ethereum account') &&
-    message.includes('URI') &&
-    message.includes('Nonce') &&
-    message.includes('Version') &&
-    message.includes('Issued At') &&
-    message.toLowerCase().includes(userStore.walletAddress.toLowerCase())
-  )
-}
-
-function getPermissionText(method, request) {
-  const { params } = request
-  let param: any
-  if (params instanceof Array && params[0]) {
-    param = params[0]
-  }
-  let response = 'performing an action'
-  switch (method) {
-    case 'wallet_addEthereumChain':
-      response = param?.chainName
-        ? `adding chain - ${param?.chainName}`
-        : 'adding a chain'
-      break
-    case 'wallet_switchEthereumChain':
-      response = param?.chainName
-        ? `switching chain - ${param?.chainName}`
-        : 'switching a chain'
-      break
-    case 'wallet_watchAsset':
-      response = 'adding a token'
-      break
-    case 'eth_sendTransaction':
-      response = 'sending a transaction'
-      break
-    case 'personal_sign':
-      if (isSiweMessage(param)) {
-        response = 'log in'
-      } else {
-        response = 'signing a message'
-      }
-      break
-    case 'eth_sign':
-      response = 'signing a message'
-      break
-    case 'eth_signTypedData_v4':
-      response = 'signing typed data'
-      break
-    case 'eth_decrypt':
-      response = 'decrypting data'
-      break
-    case 'eth_signTransaction':
-      response = 'signing a transaction'
-      break
-    default:
-      response = 'performing an action'
-      break
-  }
-  return response
-}
 
 async function onViewDetails() {
   const c = await parentConnectionStore.parentConnection?.promise
@@ -107,10 +32,6 @@ async function onViewDetails() {
       : (appStore.expandWallet = false)
   }
 }
-
-function isDeprecatedMethod(method) {
-  return method === 'eth_sign'
-}
 </script>
 
 <template>
@@ -118,19 +39,18 @@ function isDeprecatedMethod(method) {
     <div class="flex flex-col gap-1">
       <div class="flex items-center justify-center">
         <h1 class="m-0 font-bold text-lg capitalize">
-          {{ getTitle(methodAndAction[request.request.method]) }}
+          {{ props.title }}
         </h1>
       </div>
       <p class="text-sm text-center">
-        {{ appStore.name }} requests your permission for
-        {{ getPermissionText(request.request.method, request.request) }}.
+        {{ appStore.name }} requests your permission for {{ props.permission }}.
         <button class="font-bold" @click.stop="onViewDetails">
           Learn More
         </button>
       </p>
       <span
-        v-if="isDeprecatedMethod(request.request.method)"
-        class="text-xs text-yellow-100 font-medium"
+        v-if="props.deprecated"
+        class="text-xs text-yellow-100 font-medium text-center w-full"
         >WARNING: This is a deprecated method. Sign with caution.</span
       >
     </div>
@@ -157,7 +77,7 @@ function isDeprecatedMethod(method) {
       >
         <button
           class="btn-tertiary text-sm font-bold"
-          @click.stop="requestStore.skipRequest(request.request.id)"
+          @click.stop="requestStore.skipRequest(props.request.request.id)"
         >
           Do this later
         </button>
