+++ src/utils/method.ts
@@ -1,9 +1,9 @@
 export const methodAndAction = {
-  eth_sign: 'Signature',
-  personal_sign: 'Signature',
-  eth_signTypedData_v4: 'Signature',
+  eth_sign: 'Sign Message (Deprecated)',
+  personal_sign: 'Sign Message',
+  eth_signTypedData_v4: 'Sign Message',
   eth_sendTransaction: 'Send Transaction',
-  eth_signTransaction: 'Transaction Signature',
+  eth_signTransaction: 'Sign Transaction',
   eth_decrypt: 'Request Decryption',
   wallet_addEthereumChain: 'Add Chain',
   wallet_switchEthereumChain: 'Switch Chain',
